product update form separated into components
